Add tag to test that requires SSHD
The ability to do IP-based allow-listing for SSH depends heavily on the
deployment architecture.  It does *not* work out of the box with the
Cloud-Native-Hybrid deployment (one of the GET models, as used by
Dedicated) due to the use of native ingress-nginx.  It could be made to
work with gitlab-sshd and the PROXY protocol, but that is non-trivial
and not necessarily always the desired architecture.

This tag allows this test to be specifically excluded when testing
deployed instances in such architectures (e.g. for Dedicated)

This tag was added to another test with the same characteristics in a
prior MR (https://gitlab.com/gitlab-org/gitlab/-/merge_requests/93148)
but this one was (inadvertently) missed.  The key identifier for the
same need is the "except" for "review-qa-*" jobs.