spec/object_pool_service: Don't recreate object pool's source repo
The ObjectPoolService spec creates an object pool as well as a project
repository that we use as the pool's member. Both are created separately
via factory functions though, which is quite pointless given that the
pool repository factory already sets up a pool member repository for us.

Fix this duplicated setup by reusing the source project as created b the
pool repository factory.