spec/object_pool: Don't assume `#fetch` inits the pool
In commit 5dbda9650f32 (spec/object_pool_service: Don't assume `#fetch`
inits the pool, 2022-11-02), we have changed the ObjectPoolService specs
to stop relying on the `FetchIntoObjectPool()` RPC to initialize the
pool repository. This is done in preparation for an upstream change in
Gitaly to drop that functionality to have clear semantics around this
RPC and a proper lifecycle of object pools.

When writing that commit though I didn't realize that the ObjectPool
specs basically have the same dependency on this behaviour. Adjust these
specs in the same manner so that Gitaly can change behaviour of this RPC
without downstream breakage.