Disable RSpec thread safety switch
This is known to cause issues with Ruby 3,
which changed lock semantics to be per-fiber.

When testing code that uses co-routines, then
even in the absence of multiple threads this
can lead to deadlocks.

I am not aware of any reasons why we would want
to utilize multiple threads in test code, so
we should set this to false to prevent RSpec
from creating mutexed around memoizes blocks
(subject, let).

Links:
- https://github.com/rspec/rspec-core/issues/2064
- https://gitlab.com/gitlab-org/gitlab/-/merge_requests/102625