Add SLIs for Loose foreign keys cleanup
In order to get better visibility into whether or not our loose foreign
key cleanup is proceeding this MR adds SLI tracking of every time we run
the cleanup process. We consider `success` to mean that we managed to
cleanup everything within the limits of the worker process to help gain
visibility into whether the worker is regularly falling behind on the
work (thus adding delays to cleanup).

This is related to https://gitlab.com/gitlab-org/gitlab/-/issues/360501