Add partial index for `security_scans` on `id` for non-purged records
This index will fix the performance issues we are experiencing for the
batch query we run to purge security scans.

Also, with this change, we are now setting the same `created_at` date
for all the security scans to purge the scans of a pipeline at the same
time.

Changelog: fixed