Removes unused code related to the Agent metrics
The GitLab Agent Server (KAS) is the only component which uses this
internal API. In the current version of KAS it does not send
param[:gitops_sync_count] and params[:k8s_api_proxy_request] anymore.

We've kept these around while we were transitioning the KAS version to
change the params to use params[:counters]. It's safe to remove this
code now.