Squelch "Can't verify CSRF token authenticity" messages
`Gitlab::RequestForgeryProtection` is used with the API and OmniAuth
controllers. It is used to check whether a POST request is `verified?`
by checking the `X-Csrf-Token`. If there is no token, or if the token
does not decrypt properly, Rails logs a warning message, `Can't verify
CSRF token authenticity`.

Most of the time the message is not useful since with the API
`X-Csrf-Token` is only one of many ways that a user can
authenticate. On GitLab.com, it represents 30-40% of our total Rails
logs.

This commit redirects the log output to /dev/null. Controllers that
use `protect_from_forgery` alone will still generate this warning.

Changelog: changed