SAST Docs: Point to analyzers instead of upstream projects
This approach makes it clearer that we are packaging and supporting
analyzers, not using open-source projects directly off-the-shelf.

Adding "with GitLab-managed rules" clarifies that Semgrep/r2c are
not responsible for rules. We have seen rule issues reported upstream
in some cases and wish to both clarify the value GitLab provides and
reduce the effort Semgrep contributors spend redirecting to GitLab.