Add feature specs guidelines for FE
We currently do not have guidelines to
get started with feature specs as a frontend
engineer, which is often an area with friction
given that they use Ruby and Capybara, which is
outside of the usual JS expertise. This commit
adds said documentation inside the frontend
page for testing standards.