Verify method in GitlabShellWorker before using `public_send`
There are only a few valid methods that this could be calling on
Gitlab::Shell, so we explicitly list them and check in that list before
calling `public_send`.

This prevents the GitlabShellWorker being used for arbitrary code
execution if a Redis injection flaw exists - as `class_eval` is a method
on every object, `send` can be used to run literally anything.