Verify method in MailScheduler::NotificationServiceWorker
This worker is very generic, and it can call any method on
NotificationService (including those defined in EE and JH). This change
ensures that the method is defined only on those classes, not on a
superclass like Object.

This prevents the GitlabShellWorker being used for arbitrary code
execution if a Redis injection flaw exists - as `class_eval` is a method
on every object, `send` can be used to run literally anything. This
worker was also protected from that as it does not allow string
arguments.