Fix artifacts available for terraform template latest
Due to SAST-IaC creating pipelines that are tagged with merge_request.
There were then two pipelines being created with this template.
This then causes the terraform artifacts not being available.

This patch matches the workflow of SAST-IaC with Terraform
so that only one pipeline run is created for an MR. It
includes a unit test to show this.

It leaves the `deploy` job to run only on the default branch.

Solves: https://gitlab.com/gitlab-org/gitlab/-/issues/376015

Changelog: fixed