Declare `cleanup` stage in Terraform ci templates
This change adds the *by-default* unused `cleanup` stage to the 
Terraform CI templates.

The reason for this is that the base templates, 
actually prepare a `.terraform:destroy` job which is assigned to this
very `cleanup` stage.

With this change a user can easily enable that destroy job by 
configuring the pipeline like this:

```yaml
include:
  - template: Terraform.latest.gitlab-ci.yml

destroy:
  extends: .terraform:destroy
```

Before this change you would get an error that the *chosen job stage* is
not supported.