Add Package URL parser
Add a parser for package URLs (https://github.com/package-url/purl-spec)
which is based on https://github.com/package-url/packageurl-ruby.
package-url ruby does not have an active community and also uses pattern
matching which is experimental in ruby 2. So, we will copy-paste the
code and make modifications instead of including it as a gem dependency.