Add button event to detect when form is submitted
Add to all webhook forms on index / edit pages for both
project and group levels. Admin hooks use a separate
thus will be handled later.

When the button is clicked (and the form is submitted),
we call validations on the URL and URL mask items. If
there are errors, we stop the submission and scroll to
the URL section of the page.