Add active_migrations_distinct_on_table scope
To be used for fetching multiple batched migrations for execution in
parallel.

Update `active_migration` to use this new scope, with `limit: 1`.

When we switch to execution in parallel we are goign to use just the new
scope, controlling the limit with feature flag or application setting.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/372316.