Add more logs for CI config external files V2
This is the second attempt because the last one got reverted because
of a missing test case.

We have lots of lazy loading in CI config external files. And this
makes it difficult to detect the slowness. In this commit, we are adding
logging for each expensive calls.

Changelog: other