Document that a table can be dropped in a single MR
This previous documentation insisted on separating in 2 MRs but we
couldn't think of any reason this was necessary apart from slightly
derisking if it turns out that the application still needed the code.
Practically there is no reason dropping the table can't happen in the
same MR in a `post_migrate`.

This also doesn't conflict with our goal of not doing critical
migrations in `post_migrate` as these only remove stuff so it wouldn't
actually be a problem if it was skipped.