Fix Group SAML mismatched extern_uid redirect loop
If a user is signed in to GitLab locally, has an existing Group
SAML identity, and tries to sign in to a group with SSO, they
were stuck in a redirect loop. This fixes the loop and instead
updates the identity with the new extern_uid when the email
addresses from SAML and on the user match.

Changelog: fixed
EE: true