Test sub entries of CI schema
For each JSON schema that we validate, we
go through each key individually and make
sure that it is valid or invalid
as a sub-schema. This allow us to
have a single file with multiple
invalid examples, and each of them
indivudually has to fail for the test
to pass.