Ignores ingress fetch failure for k8s less than 1.19
This feature relates to the Deploy Boards. A feature which is already
deprecated but still being used by some users. As of today, we only
support Kubernetes version from 1.23 and above. But there are still some
old clusters with versions lower than 1.19 that are connected to GitLab.
Those will fail to load ingress resources as they the ingress api
endpoint had a breaking change since then. Since we only support 1.23
and above, we will ignore ingress fetching errors below that. But
actually it could still work from at least 1.19.

To check cluster version we could not use kubeclient, as it does not
support querying the /version Kubernets endpoint. So we are using
Gitlab::HTTP instead.

Relates to: https://gitlab.com/gitlab-org/gitlab/-/issues/371249

Changelog: fixed