Fix flaky test ordering problem in transaction_leaking_spec
I tried a bunch of different ways to fix the test ordering problem
described in https://gitlab.com/gitlab-org/gitlab/-/issues/382439 .
It seemed that it was not enough to reset the `for_model` method in
`after do` which makes sense because these stubs are already reset
between tests. I needed to also call `Setup` again on the model because
the incorrect host configuration was memoized on the model which is not
reset between tests. Then I found this alone was not sufficient beause I
noticed that the test is using `for_model(ApplicationRecord)` which is
not actually ever being called in the initializer in normal code
execution. It seems it's meant to be `ActiveRecord::Base` and we do not
correctly clear whatever is cached between tests unless we use that
consistently.