Implement AccessLevel User and Group types with limited fields
Previously we have been returning core user and group types in the
GraphQL API endpoint under the
branchRules.branchProtection.{Push,Merge,Unprotect}AccessLevels nodes.
A lot of this data is not required to display branch rules and is
challenging to preload as the associations could be nested infinitely.
This change introduces new User and Group types that expose a subset of
fields. This will allow us to preload this date easier.

EE: true
Changelog: changed