Adjust error check to less strict regexp
In the MR https://gitlab.com/gitlab-org/gitaly/-/merge_requests/5029
the errors creation was a bit refactored to exclude the RPC name as
it is redundant information. As it will became a cause of the
GitLab tests failure after merge we make an error text check less strict.
It allows us to remove redundant error prefix, and also keeps the
verification of the cause in place.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/4471