Improve epic swimlanes performance
This aims to improve the epic swimlanes performance on issue boards.
It rewrites the query using a CTE for all the epics a user has access to
and an EXISTS query where it JOINs the issues with the allowed epics.

We're also sorting the allowed epics as it's then in the order expected
by the resolver.