Change feature flag in JC OAuth setting controller
This changes the feature flag used to control the exposure of the
jira_connect_application_key setting. When GitLab.com is used as a
proxy for the GitLab.com for Jira Cloud app, it needs to fetch the
OAuth application ID from a self-managed instance.
jira_connect_oauth_self_managed_setting feature flag
will be enabled by default and rolled out to self-managed users.

This prepared self-managed users for the new feature and they will be
able to use it as soon as the jira_connect_oauth_self_managed
feature is enabled on GitLab.com