Use the package-url test suite
Use the suite of tests provided alongside the package-url specification
at https://github.com/package-url/purl-spec/blob/56fb237fbe0368e5e3c5df90bcfc516d51d95944/test-suite-data.json

This ensures that the parser conforms to the specification. As a result
of this change, I've had to update parts of the parser which were not
conforming to the specification.