Document how to fix PDF views with LFS and proxy_download disabled
When LFS storage is configured with object storage and
`proxy_download` is set to `false`, users attempting to view PDF files
uploaded to LFS may see:

An error occurred while loading the file. Please try again later.

This can happen because the object storage provider rejects the PDF.js
frontend request to load the PDF file since the GitLab domain is not
allowed due to Cross-Origin Resource Sharing (CORS) restrictions. This
commit documents that it is necessary to configure S3 CORS headers to
make this work.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/248100

Changelog: fixed