Properly change the URL when changing MR tabs
The fix for the problem is just in moving the
`.setCurrentAction` call into the `tabShown` function,
which should guarantee we're updating the URL every
time the tab changes.

However, the back button is still an issue here, because
the `popstate` handler is doing special logic outside of
the `tabShown` logic. By EXPECTING the popped entry to
have a `state` object (which we would have assigned).
For some requests (especially first loads), this is not true.

To "standardize" everything, the `popstate` handler now
just defers to the `tabShown` function with a simple
"extraction" of the action from the URL.

Changelog: fixed