Fix client logging for job users
When we have a job user, `user_id` will exist (as will `job_user`), but
`user` won't. This meant we were showing IPs for these, even though we
were logging `user_id`.

To fix this, and allow for runner ID to still take precedence, we make
`client_id` depend on the previous value of `client_id` (if any) to
account for the context stack.