Potential fix for flaky spec in resolve discussions spec
The `user_resolves_diff_notes_and_discussions_resolve_spec.rb` feature
may be flaky because the test attempts to:

1. Resolve a thread
2. Reply to a thread to unresolve it
3. Confirm that the thread has been unresolved

There may be a race condition between steps 1 and 2. The test should
wait for the discussion to be resolved before attempting to write a
new note.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/341546