Update expired certs
The affected spec was supposed to be removed after 15.0 because
the underlying feature was supposed to be removed. However, it was
postponed to 16.0 and therefore, this spec now caused problems because
the used certs [expired a few days
ago](https://gitlab.com/gitlab-org/gitlab/-/issues/347425#note_758005033).

This change is **not** a permanent fix as it'll break again next summer
at exactly `Aug 23 19:42:10 2023 GMT` - because the
`leaf_certificate.pem` will expire again.

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/383197