Stop using allow_disk_access in projects transfer service specs
These specs were asserting the absolute disk path as used by gitaly to
determine the repository location. This will not be possible when using
gitaly cluster as there is no way to determine which node the repository
will be on and hence the filesystem path. Instead use the relative path.
This is the path of the repository, minus the gitaly specific root.