Remove flag exemption for webhook rate limits
This removes a flag that had been added to webhook rate limits in order
to provide an exemption for a single customer. The exemption was in
order to rollout the rate limit feature while causing minimal disruption
to existing customers.

The customer is now in a position where we can remove the exemption
permanently and be subject to webhook rate limits.

https://gitlab.com/gitlab-org/gitlab/-/issues/366081
https://gitlab.com/gitlab-org/gitlab/-/issues/365605
https://gitlab.com/gitlab-org/gitlab/-/issues/366084

Changelog: changed