Fix helper to only list conflicts when source/target branch exists
When source/target branch is deleted the
`MergeRequests::Conflicts::ListService` call will fail because
it requires for both source/target branch to exists. This can
result to the MR changes tab not loading when there are
conflicts.

To fix that, we check if source/target branch exists before making
the call.

The call for `DiffHelper#conflicts` is behind the
`display_merge_conflicts_in_diff` feature flag.