Fix group transfer not transfering groups
With the feature flag for cross-group hierarchy linking enabled, we
allow epics to move with the group. However, when the feature flag is
disabled we skipped the call to the actual move of the group to its new
parent.

This fixes the issue by always calling `super` and only "nullify" lost
group parent if the feature flag is disabled.

Changelog: fixed
EE: true