In 53fb7663 (Support structured error for UserCommitFiles, 2022-10-24) we updated #user_commit_files to accept structured errors. As part of this change we updated the error messages when a structured error is returned to include the file path. However, a number of tests check for specific error strings and fail in this scenario.
To resolve this, make structured error messages be exactly the same as
the error strings previously returned by Gitaly.

We cannot simply adjust the tests to use the new errors as currently
they are behind a feature flag in Gitaly, and tests will need to support
both sets of errors. Once Gitaly has removed the feature flag for
structured errors it will be possible to adjust these errors as-needed.

Changelog: fixed