Configures agentk to know when to talk to canary
QA tests run against canary when the environment var
QA_COOKIES="gitlab_canary=true" is set. Up to now, agentk was always
talking to the non-canary infrastructure even if this cookie was set.

With this change we tell kas to add a canary cookie to its requests so
that its requests get redirected to the KAS canary server.

Issue: https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2557