Fix non menu-item button styles in dropdowns
We previously styled any link or button inside a dropdown
as if it were a menu item. For some forms in dropdowns like
Apply Suggestions this meant the button got some unwanted
styles.

Ideally we should just remove this rule since it's a lot
of selectors for what should just be dropdown-menu-item or
something. But that is much more likely to break things
so will consider it later

Changelog: fixed