Verify method name in MailScheduler::NotificationServiceWorker
This worker uses `public_send` to call one of a large number of methods
on a NotificationService instance. Previously, we didn't verify the
method name, which meant that this could be used as a vector to escalate
an existing vulnerability (for instance, Redis command injection) into
arbitrary command execution (for instance, via `instance_eval`).

We now do verify that the method is defined directly on
NotificationService, or, where present, EE::NotificationService or
JH::NotificationService.

The feature flag was used to do a graceful rollout to production in case
there were any issues. There aren't, so we can always verify the method
name we're about to call in MailScheduler::NotificationServiceWorker.

Changelog: other