Do not save PipelineMetadata if name is blank
If the pipeline name is blank after substituting variables, do not
save the name.

If the name is saved, it prevents creation of the pipeline since the
empty name fails validation.

Not saving the empty name allows pipeline names that don't need to
have a value for all possible pipelines.