Allow IssuePolicy.can_read_confidential to be preloaded
The IssueCollection class is very old and only used in
the IssuePolicy class. This class does not use our
redis cache for user max access on a project. This doesn't
allow the max access to be preloaded for a list of projects.
The change uses the cached issue policy reporter access check

Changelog: performance