Workhorse: simplify gRPC call metadata propagation
When Workhorse makes a gRPC call to Gitaly, it adds call metadata
provided by Rails such as the username. The way this is currently done
is verbose and forces us to list the metadata we want to propagate at
each call site in Workhorse. This commit changes things so propagation
happens automatically.

Changelog: other