Stop using allow_disk_access in rename namespaces/projects migration specs
These specs were relying on allow_disk_access to get the physical disk
path of repositories in order to check that they exist. With gitaly
cluster it will no be possible to determine where the repository is
stored. So instead we refactor these tests to use the RepositoryExists
RPC instead.