Install two versions of Sentry Client SDK
This change installs two versions of sentry that can report errors to
two Sentry instances respectively.

Reduces the bundle size by having the old and new Sentry SDKs in
separate bundles and only loading one of them depending on the
state of the feature flag.

A global object is used to "re-export" the current Sentry so it can used
by the clients.

Changelog: changed