Remove custom CORS controller for JiraConnect
This goes back to use Rack Cors instead of the custom cors
controller for jira_connect routes.

The customer controller was added to be able to add origins dynamically
based on the jira_connect_proxy_url setting. I learned that this can be
done by passing a block to origins using Rack Cors.

The custom controller also had the problem that the OPTIONS
request was intercepted by Rack Cors and never reached the controller.

Changelog: fixed