Add migration for jobs, update schema_migrations
Since Clusters::Applications has been deprecated, we are in the process
of cleaning up lingering code around this part of the application. This
post migration removes job class instances that may be lingering in in
queues, in an effort to ensure that the eventual removal of the job
classes themselves (this will take place in a future milestone) is safe
to do.

The following job classes are affected:
ClusterConfigureIstioWorker
ClusterInstallAppWorker
ClusterPatchAppWorker
ClusterUpdateAppWorker
ClusterUpgradeAppWorker
ClusterWaitForAppInstallationWorker
ClusterWaitForAppUpdateWorker
ClusterWaitForIngressIpAddressWorker

Changelog: other