Refactor SCA::LicenseCompliance to use License Scanning Service
We're establishing a contract that the license scanning service
should be used to retrievee the license scanning report. For
the time being this will proxy all of the calls so that it
fetches the json artifact from the latest pipeline. In the
future, the implementation will change so that it retrieves
the projects SBoM components and fetches their respective
licenses from the LicenseDB. These licenses will then be
used to create the license scanning report on the fly so that
the service contract remains fulfilled.