Improve Dast site profile creation error message
This Merge Request adds a size validation
for the password and request headers
parameters in the dast site profile
creation.

These attributes are maskable.
The maskable module has a minimum size
validation of 8 characters. This Merge
Request repeats this validation on the
service to make the error clear for our
users.

Changelog: fixed
EE: true