Relax UX reviewer choice requirements
This change relaxes the requirement to use the Reviewer Roulette as the
sole way to find a UX reviewer.

There can be several reasons for an author to not be able to ask for a
review from a randomly assigned reviewer. This change aims to allow
authors to override the recommendation while also advising them to
follow it.