Allow commits to be signed with SSH keys
You can now use SSH keys for signing commits. Signing commits provides
others with confidence that a commit or tag was authored by you. In the
UI, signed commits show as "verified".

Previously, it was only possible to sign commits with GPG keys or with
X.509 certificates, requiring you to manage different key types for
authentication and for signing. To enable signing with SSH keys go to
your avatar in GitLab Edit profile > SSH Keys and add a key for signing
commits and for authenticating against GitLab. If you already had an SSH
key for authentication setup in your profile, it will automatically be
used for signing commits unless you also had a GPG key or X.509
certificate in your profile to sign commits.

Changelog: added