Only retry on network error on suggested reviewers
We would like to track only network errors when calling the suggested
reviewers client. These types of errors will deplete the error budget.

Other application errors that are deep in the stack, that are not
network-related are not retriable but will be tracked and reported
internally in the suggested reviewers monitoring infrastructure.