Add specific state name URI requirements
This change allows everything in a Terraform state name except slashes.
It's the same requirement as for namespaces and projects.

This solves a common confusion that when a state name contained a dot it
resulted in a `404`.

Closes: https://gitlab.com/gitlab-org/gitlab/-/issues/219460

Changelog: added
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/219460
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/105674