Protect environments with multiple approval rules
Changes to the UI behind the feature flag multiple_approval_rules_fe
that allow users to protect environments with multiple approval rules, a
stronger variant. Once complete and the feature flag is rolled out, the
unified approval rule creation UI will be removed.

Multiple approval rules behave much like approval rules for merge
requests: a group, user, and/or access level can be separately
configured to be required to approve a deployment before it is run.