Set a weak password threshold for "forbidden substrings"
`Security::WeakPasswords` flags a password as weak when it contains a
user's username, name, email, or a "substring" of one of those. E.g.
for a user with a full name "Nick Malcolm" if either "Nick", "Malcolm"
or "Nick Malcolm" appeared in the password, it would be weak.

However in long passwords it was possible that those forbidden
substrings would legitimately occur in a truly random password. This in
fact did occur in the randomized GitLab specs
(See https://gitlab.com/gitlab-org/gitlab/-/issues/384336). This MR,
therefore, skips the "weak substring" checks for passwords 64
characters or longer.

Passwords are still checked against the static weak password list
regardless of length.

- https://gitlab.com/gitlab-org/gitlab/-/issues/384336