Deduplicate increments during a refresh
During a refresh, BufferedCounter keeps track of
existing increment and decrements on specified ref.
It deduplicates each change based on the historical change,
ensuring a net change to the counter that reflects
the current state, regardless of the order that each
increment or decrement comes in.

Adds a new state on BuildArtifactsSizeRefresh
`finalizing` which delays execution of `#finalize!`.
This is to ensure all decrements are deduplicated.

In `#finalize!`, it will move the current value
in the refresh counter key into the counter key.
This will eventually be committed into database
as per normal.

Changelog: fixed