Remember the email after registration errors
When registering an account and encountering an error, the email field
would not be remembered / be reset to an empty string. This MR
fixes this UX bug.

The root cause was that the form assumed that if an email ever needed
to be populated in the form, that it'd come from a value `@invite_email`
which is only used during Invitation flows.

https://gitlab.com/gitlab-org/gitlab/-/issues/384537