Refactor tests to use `createAlert` part 1
`app/assets/javascripts/filtered_search/visual_token_value.js` and
`app/assets/javascripts/filtered_searc h/filtered_search_manager.js`
were migrated to use `createAlert` in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/98949.

However, the test were not updated to use `createAlert` because they
didn't failed. The reason for that is because the test were asserted to
not be invoked.