Process Slack interactivity events in-request
Responses by GitLab to Slack interactivity request must happen within
3 seconds when the interactivity request was a modal interaction.
https://api.slack.com/interactivity/handling#modal_responses

We currently handle Slack interactivity requests by queuing to a worker.

The worker must be marked as `worker_has_external_dependencies!` but
this also means that we must mark it as `urgency :low`.

As we want to ensure that we respond to modal events in time, this
change makes the endpoint that receives Slack interactivity requests
respond in-request.

The endpoint has been marked as `urgency: :low` to reflect the Apdex
of an endpoint that only processes non-user interactions.