Fix flaky project destroy service spec
The `UpdateCachedMetadataWorker` gets called with an array of sliced
ids. e.g. `[[1,2], [3]]`. As we can't be sure of the ordering of the IDs
we had a flaky spec.
Since the array is sliced, we were not able to just use `match_array`
and therefore opted to check separately if the ids got sliced into two
arrays and contain the right IDs.