Add subscription for user notes on work items
Adding notes related subscriptions on note create, update, delete.

On note create, update a NoteType is returned containing the note data
On note delete a DeletedNoteType is returned that contains note ID,
note's discussion IDs and a last note boolean flag that signals
if that is the last note in the discussion or not.

Changelog: added