Fix update reviewers request when none is assigned
This fixes an update error when the front end attempts to update merge
request reviewers without actually assigning anyone. The merge request
model does not know how to handle `suggested_reviewer_ids` param and
causes the request to err with 500.

We would like to ensure the `suggested_reviewer_ids` param is always
removed from the params before passing it to the model for persistence.