Ensure TooManyIps exception returns 403 forbidden for GraphQL
When a user sign-ins from too many different IP addresses, GitLab
may raise a Gitlab::Auth::TooManyIps exception. Most portions of
the app rescue that error and respond in a meaningful way depending
on whether the response is in a browser, a JSON response to the
REST API, etc. However, until this point the GraphQL API was
responding with a 500 internal server error.

Changelog: fixed