Enable pipeline_tabs_vue by default
This commit removed the FF `pipeline_tabs_vue`
which in effect will enable it by default for all
instances. It is currently at 100% on .com. There is
also a small bug fix found by re-enabling
the specs for a show file. We need to not render
the pipeline tabs when htere is a yaml error in
the config. This was not the case with the new
behaviour probably due to an indentation typo.

Changelog: changed