Clarify permissions required for vulnerability APIs
The fact that we delegate permissions to the project for vulnerabilities
is an implementation detail, and it is confusing for users as, from an
outsider perspective, all permissions come from projects. Based on my
testing, trying to request a vulnerability which belongs to a project
that you aren't a member of results in a 403. Trying to request a
vulnerability which doesn't exist results in a 404. I've decided not to
mention this as it is the default behavior described in
`doc/api/index.md#status-codes`.