Return HTTP 400 for multipart header without boundary
If the client passes a multipart/form-data content type without a
boundary, Workhorse would return an HTTP 500 status code (Internal
Server Error). This should be status 400 (Bad Request). This helps the
user because it tells them the problem is on their end (in the
request) and not on the GitLab server.

Fixes https://gitlab.com/gitlab-org/gitlab/-/issues/385112.

Changelog: fixed