Add `Remember me` in the SAML workflow
We added a Remember me checkbox to the SAML authorization workflow, both
in the autorization and sign in steps.

Changes in the UI: we replaced several emphasised words by link.

Added `autocomplete="off"` to the `Remember me` checkbox

This is the problem we want to solve: The user could be presented with a
`Remember me` checkbox checked but the URL of the button didn't have the
`remember_me` query parameter. The following sequence of steps reproduce
this problem:

1. click on the `Remember me` checkbox
2. press the button (URL of the button correctly contains the parameter)
3. go back (with the browser arrows)
4. the `Remember me` is checked but the URL of the button doesn't
   contain the parameter)

By adding `autocomplete="off"` attribute, when taking step 3, it
results in a checkbox that is always unchecked.

Fix failing feature tests:

* Added `:js` to enable JavaScript
* Modified a little tests where the JavaScript was used to automatically
  re-authenticates using `js-auto-redirect-to-provider`.
* Added `with_omniauth_full_host { example.run }` to emulate the
  redirection from the SAML provider.

Changelog: changed
EE: true