Add steps to check whether onboarding is finished
It checks to see if the Jitsu key exists, if it doesn't then we're still
in onboarding.

Then it tries to query Cube, if cube fails or the returned count is 0
then we're still in onboarding.

If cube returns a count greater than 0 then we have data and we're onto
listing dashboards!

Signed-off-by: Robert Hunt <rhunt@gitlab.com>