Change relative position assigns for epic boards
When an epic gets moved within a board and we need to assign the
relative_position for all the epics in the list, we previously set the
relative position with a multiple of the starting position.

For large numbers as the starting positions, this can lead to issues as
we arrive at the MAX Integer value rather quickly.

To overcome the problem, we now use a constant factor (the
`IDEAL_DISTANCE` constant) as the distance between each relative
position.

Changelog: fixed
EE: true